requirements.txt
aiosmtpd
ruamel.yaml
click
